+++ packages/algob-web/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: [
+    "../../.eslintrc.js"
+  ],
+  parserOptions: {
+    project: "./test/tsconfig.json"
+  }
+}
