+++ packages/algob-web/README.md
@@ -1,0 +1,1 @@
+# Algob-web
