+++ packages/algob-web/src/index.ts
@@ -1,0 +1,1 @@
+export * as types from "./types";
