+++ packages/algob-web/src/lib/constants.ts
@@ -1,0 +1,3 @@
+export const MIN_UINT64 = 0n;
+export const MAX_UINT64 = 0xFFFFFFFFFFFFFFFFn;
+export const reDigit = /^\d+$/;
