+++ packages/algob-web/tsconfig.json
@@ -5,6 +5,6 @@
   },
   "include": [
     "src",
-    "../../@types", "../algob-web/src/types-input.ts"
+    "../../@types"
   ]
 }
